try to manage videos which frame rate is not an integer.
- In QqWidget::QqWidget, modified the FPS to be changed in percentage.
changed actualFps and normalFps to be taken as a float.

- In FPS:: , modified to use float instead of int.

- In oggmux_add_audio, limit the ratio to +- 10% resampling, which
is a lot ..... just in case of big drift.

+++ src/fps.cpp
@@ -58,7 +58,7 @@ FPS::~FPS() {
     error("error destroying POSIX thread feed attribute");
   
 }
-void FPS::init(int rate) {
+void FPS::init(float rate) {
 
 
   this->set(25);
@@ -110,12 +110,12 @@ void FPS::calc() {
   
 }
 
-int FPS::get() {
+float FPS::get() {
   return (_fps ? fpsd.sum / fpsd.n : 0 );
 }
 
-int FPS::set(int rate) {
-  func("FPS set to %u",rate);
+float FPS::set(float rate) {
+  func("FPS set to %f",rate);
   if (rate < 0) // invalid
     return fps_old;
